setting up task routers
