+++ src/index.js
@@ -1,8 +1,8 @@
 const express = require("express");
 require("./db/mongoose");
-const User = require("./models/user");
-const Task = require("./models/task");
+
 const userRouter = require("./routers/user");
+const taskRouter = require("./routers/task");
 
 const app = express();
 const port = process.env.port || 3000;
@@ -10,75 +10,7 @@ const port = process.env.port || 3000;
 app.use(express.json()); // this configures express to automatically parse JSON into
 // object so we can access it in our request handlers
 app.use(userRouter);
-
-app.post("/tasks", async (req, res) => {
-  const task = new Task(req.body);
-  try {
-    await task.save();
-    res.status(201).send(task);
-  } catch (e) {
-    res.status(400).send(e.message);
-  }
-});
-
-app.get("/tasks", async (req, res) => {
-  try {
-    const tasks = await Task.find({});
-    res.send(tasks);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
-});
-
-app.get("/tasks/:id", async (req, res) => {
-  const _id = req.params.id;
-  try {
-    const task = await Task.findById(_id);
-    if (!task) {
-      return res.status(404).send(); // 404:not found
-    }
-    res.send(task);
-  } catch (e) {
-    res.status(500).send(e.message);
-  }
-});
-
-app.patch("/tasks/:id", async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ["completed", "description"];
-  const isValidOperation = updates.every(update =>
-    allowedUpdates.includes(update)
-  );
-
-  if (!isValidOperation) {
-    return res.status(400).send({ Error: "Invalid updates!" });
-  }
-
-  try {
-    const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true
-    });
-    if (!task) {
-      res.status(404).send();
-    }
-    res.send(task);
-  } catch (e) {
-    res.status(400).send(e.message);
-  }
-});
-
-app.delete("/tasks/:id", async (req, res) => {
-  const task = await Task.findByIdAndDelete(req.params.id);
-  try {
-    if (!task) {
-      return res.status(404).send();
-    }
-    res.send(task);
-  } catch (e) {
-    res.status(500).send();
-  }
-});
+app.use(taskRouter);
 
 app.listen(port, () => {
   console.log(`Server is up on port ${port}`);
